chore(frontend): rename Publications chart to match the camel-case naming convention
