+++ frontend/src/components/charts/Publications.tsx
